feat(graph): implement cross-program dependency graph (Closes #26)
